feat: Update plugins and improve key mappings
This commit includes changes to the plugin configuration and key mappings.The 'lazy-lock.json' file was added to track plugin versions. The'better-escape.lua' plugin was added for improved key escape functionality.The 'completion.lua' file was updated to include an optimization prompt andimprove key mappings. The 'editor.lua' file was updated to include the'better-escape.nvim' plugin. The 'dots.tutor' file was removed as it is nolonger needed.